+++ internal/model/commit.go
@@ -1,0 +1,13 @@
+package model
+
+import "time"
+
+type RepoCommit struct {
+	Scope    string
+	Hash     string
+	Author   string
+	When     time.Time
+	Subject  string
+	IsRoot   bool
+	RepoPath string
+}